tools/gcp-filestore-monitor/gcp-filestore-monitor/requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cloudevents==1.2.0
colorama==0.4.4
deprecation==2.1.0
Flask==2.0.2
functions-framework==3.0.0
google-api-core==2.3.0
google-auth==2.3.3
google-cloud==0.34.0
google-cloud-access-context-manager==0.1.9
google-cloud-common==1.0.0
google-cloud-core==2.2.1
google-cloud-filestore==1.0.0
google-cloud-monitoring==2.8.0
google-cloud-org-policy==1.2.1
google-cloud-os-config==1.9.0
google-cloud-resource-manager==1.3.3
google-cloud-storage==1.44.0
google-cloud-tasks==2.7.2
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpc-google-iam-v1==0.12.3
grpcio==1.42.0
grpcio-status==1.42.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.4
packaging==21.3
prometheus-client==0.12.0
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dotenv==0.19.2
requests==2.26.0
rsa==4.8
six==1.16.0
urllib3==1.26.7
watchdog==1.0.2
Werkzeug==2.0.2
